fix(toggle-segment)!: tweak focus to be :focus-visible
Tweak ToggleSegment component and corresponding css to make use
of the :focus-visible state of the hidden input rather than
:focus-within (on wrapping label). Done because focus-visible only
triggers by user agent definition. E.g. it will not trigger on
click in browsers, but will trigger on keyboard navigation.

Remove a few obsolete css properties.

BREAKING CHANGE: The change might be breaking in cases where one
has selected elements in custom css, as the semantics of the
component are changed, or one has made use of custom html markup.